requirements.txt
groq
streamlit
python-dotenv 
python-docx
requests